rename : view폴더 생성
