+++ src/view/OutputView.js
@@ -1,5 +1,5 @@
-const { OUTPUT } = require("./constants/message");
-const { print } = require("./utils/utils");
+const { OUTPUT } = require("../constants/message");
+const { print } = require("../utils/utils");
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
  */
